--- fe/src/routes/IssueList.tsx
@@ -1,11 +1,0 @@
-import Header from "../components/Header";
-
-function IssueList() {
-	return (
-		<div className="h-[95%] w-[85%] ">
-			<Header />
-		</div>
-	);
-}
-
-export default IssueList;
